♻️ Extract public/con.go from conn.go
