+++ public/conn.go
@@ -1,0 +1,43 @@
+package public
+
+import (
+	"database/sql/driver"
+	"errors"
+	"log/slog"
+)
+
+func ConnExecContextErrorHandler(driverName string) func(err error) (bool, []slog.Attr) {
+	switch driverName {
+	case DriverNameMysql:
+		return func(err error) (bool, []slog.Attr) {
+			if err == nil {
+				return true, nil
+			}
+			// https://pkg.go.dev/database/sql/driver#ErrSkip
+			if errors.Is(err, driver.ErrSkip) {
+				return true, []slog.Attr{slog.Bool("skip", true)}
+			}
+			return false, nil
+		}
+	default:
+		return nil
+	}
+}
+
+func ConnQueryContextErrorHandler(driverName string) func(err error) (bool, []slog.Attr) {
+	switch driverName {
+	case DriverNameMysql:
+		return func(err error) (bool, []slog.Attr) {
+			if err == nil {
+				return true, nil
+			}
+			// https://pkg.go.dev/database/sql/driver#ErrSkip
+			if errors.Is(err, driver.ErrSkip) {
+				return true, []slog.Attr{slog.Bool("skip", true)}
+			}
+			return false, nil
+		}
+	default:
+		return nil
+	}
+}
